Group the host reply callbacks by what the reply carries
The switchyard file still held every one of the two dozen
host-callback to reply functions, which left it at 2900 lines - the
same problem the previous commit set out to fix, just moved. The
callbacks do fall into groups, though, because what they have to
pack determines their shape.

Split them along that line. pmix_server_op_replies.c takes the ones
that complete a specific client operation and carry that operation's
own payload - the fence modex, a direct-modex get, connect,
disconnect, spawn, lookup, event registration, and IOF register and
deregister. pmix_server_info_replies.c takes the eleven that hand
host-supplied results back, all of which share one shape: the host
returns an array of pmix_info_t (or a credential, or a set of device
distances), we pack it and queue it to the caller. Those eleven are
near-identical to each other by design, so keeping them side by side
is what makes the advice to diff a new one against its neighbor
actually possible.

The switchyard keeps the generic status-only replies - op_cbfunc,
op_cbfunc2 and resop_cbfunc - because a dozen of its arms hand those
out by name and they are effectively part of dispatch. That leaves
the three files at roughly 760, 1100 and 1200 lines.

The twenty callbacks that moved are named by the switchyard, so each
takes the pmix_server_ prefix and a declaration in
pmix_server_ops.h. Their inner _*cbfunc progress-thread handlers
stay static and stay adjacent to them - nothing else calls those.
Five switchyard arms had their assignment lifted out of the if
condition to stay inside the line limit; the behavior is unchanged.

Also move the host-callback banner comment out of
pmix_server_pset.c, where the previous commit had left it stranded.
